+++ R/stdEff-fun.R
@@ -265,18 +265,19 @@ xNam <- function(mod, intercept = TRUE, aliased = TRUE, list = FALSE,
 #'   numeric variable, using the link function from a generalised linear model
 #'   (GLM) fit to the variable. The transformation is generalised in the sense
 #'   that it can extend even to cases where a standard link transformation would
-#'   generate undefined values. It achieves this by using an estimate based on
+#'   produce undefined values. It achieves this by using an estimate based on
 #'   the 'working' response variable of the GLM (see below). If the error
 #'   distribution `family` is not specified (default), then it is determined
 #'   (roughly) from `x`, with `binomial(link = "logit")` used when all x <= 1
 #'   and `poisson(link = "log")` otherwise. Although the function is generally
-#'   intended for binomial or poisson variables, any variable which can be fit
-#'   using [glm()] can be supplied. One of the key purposes of `glt()` is to
-#'   allow the calculation of fully standardised effects (coefficients) for GLMs
-#'   (in which case `x` = the response variable), while it can also facilitate
-#'   the proper calculation of SEM indirect effects (see below).
+#'   intended for variables with a binomial or Poisson distribution, any
+#'   variable which can be fit using [glm()] can be supplied. One of the key
+#'   purposes of `glt()` is to allow the calculation of fully standardised
+#'   effects (coefficients) for GLMs (in which case `x` = the response
+#'   variable), while it can also facilitate the proper calculation of SEM
+#'   indirect effects (see below).
 #'
-#'   **Estimating the link transformation**
+#'   **Estimating the direct link transformation**
 #'
 #'   A key challenge in generating fully standardised effects for a GLM with a
 #'   non-gaussian link function is the difficulty in calculating appropriate
@@ -289,7 +290,7 @@ xNam <- function(mod, intercept = TRUE, aliased = TRUE, list = FALSE,
 #'   a more general variance-based method using a pseudo R-squared (Menard 2011)
 #'   – here an alternative approach is used. Where transformed values are
 #'   undefined, the function will instead return the synthetic 'working'
-#'   response from the iteratively reweighted least squares algorithm (IRLS) of
+#'   response from the iteratively reweighted least squares (IRLS) algorithm of
 #'   the GLM (McCullagh & Nelder 1989). This is reconstructed as the sum of the
 #'   linear predictor and the working residuals — with the latter comprising the
 #'   error of the model on the link scale. The advantage of this approach is
@@ -308,39 +309,40 @@ xNam <- function(mod, intercept = TRUE, aliased = TRUE, list = FALSE,
 #'   fitting procedure to calculate a 'final' working response:
 #'
 #'   1. A new GLM of the same error family is fit with the original response
-#'   variable as both predictor and response, and using a single IWLS iteration.
+#'   variable as both predictor and response, and using a single IRLS iteration.
 #'
-#'   2. The working response is extracted from this model.
+#'   2. The working response is reconstructed from this model.
 #'
-#'   3. The inverse transformation of the working response is then calculated.
+#'   3. The inverse transformation of the working response is calculated.
 #'
 #'   4. If the inverse transformation is 'effectively equal' to the original
-#'   response (tested using [all.equal()]), the working response is returned;
-#'   otherwise, the GLM is refit with the working response now as the predictor,
-#'   and steps 2-4 are repeated — each time with an additional IWLS iteration.
+#'   response (tested using [all.equal()] with default tolerance), the working
+#'   response is returned; otherwise, the GLM is refit with the working response
+#'   now as the predictor, and steps 2-4 are repeated — each time with an
+#'   additional IRLS iteration.
 #'
 #'   This approach will generate a very reasonable transformation of the
 #'   response variable, which will also be practically indistinguishable from
-#'   the direct transformation where this can be compared (see Examples). It
+#'   the direct transformation, where this can be compared (see Examples). It
 #'   also ensures that the transformed values, and hence the standard deviation,
 #'   are the same for any GLM fitting the same response (provided it uses the
 #'   same link function) — facilitating model comparisons, selection, and
 #'   averaging.
 #'
 #' @note As we often cannot directly observe the GLM response variable on the
-#'   link scale, any method estimating its values or statistics will incorporate
-#'   a degree of error. The aim should be to try to minimise this error as far
-#'   as (reasonably) possible, while also generating standardised effects whose
-#'   interpretation most closely resembles those of an ordinary linear model -
-#'   something which the current method achieves. The solution of using the
-#'   working response from the GLM to scale effects is a purely practical, but
-#'   reasonable one, and one that takes advantage of modern computing power to
-#'   minimise error through iterative model fitting. An added bonus is that the
-#'   estimated variance is constant across models fit to the same response
-#'   variable, which cannot be said of previous methods (Menard 2011). The
-#'   overall approach would be classed as 'observed-empirical' by Grace *et al.*
-#'   (2018), as it utilises model error variance (the estimated working
-#'   residuals) rather than theoretical distribution-specific variance.
+#'   link scale, any method estimating its values or statistics will be wrong to
+#'   some degree. The heuristic approach described here aims to reduce this
+#'   error as far as (reasonably) possible, while also generating standardised
+#'   effects whose interpretation most closely resembles those of the ordinary
+#'   linear model. The solution of using the working response from the GLM to
+#'   scale effects is a practical, but reasonable one, and one that takes
+#'   advantage of modern computing power to minimise error through iterative
+#'   model fitting. An added bonus is that the estimated variance is constant
+#'   across models fit to the same response variable, which cannot be said of
+#'   previous methods (Menard 2011). The overall approach would be classed as
+#'   'observed-empirical' by Grace *et al.* (2018), as it utilises model error
+#'   variance (the estimated working residuals) rather than theoretical
+#'   distribution-specific variance.
 #' @return A numeric vector of the transformed values, or an array, list of
 #'   vectors/arrays, or nested list.
 #' @references Grace, J.B., Johnson, D.J., Lefcheck, J.S. and Byrnes, J.E.K.
@@ -400,7 +402,7 @@ glt <- function(x, family = NULL, force.est = FALSE) {
       )
       i <- 0
       repeat {
-        xl <- predict(m) + resid(m, "working")
+        xl <- predict(m) + resid(m, type = "working")
         xli <- f$linkinv(xl)
         eql <- isTRUE(all.equal(x, xli, tolerance = sqrt(.Machine$double.eps),
                                 check.names = FALSE))
@@ -475,12 +477,12 @@ getY <- function(mod, data = NULL, link = FALSE, offset = FALSE,
     }
 
     # Model response
-    y <- fitted(m) + resid(m, "response")
+    y <- fitted(m) + resid(m, type = "response")
     if (!is.null(w)) y <- y[w > 0 & !is.na(w)]
     if (!is.null(o)) y <- f$linkinv(f$linkfun(y) - o)
     y[zapsmall(y) == 0] <- 0
-    a <- names(attributes(y))
-    attributes(y)[a != "names"] <- NULL
+    an <- names(attributes(y))
+    attributes(y)[an != "names"] <- NULL
 
     # Return response on original or link scale
     if (isGlm(m) && link) glt(y, family = f) else y
@@ -1224,7 +1226,9 @@ stdEff <- function(mod, weights = NULL, data = NULL, term.names = NULL,
     if (k > 0) {
 
       # Predictors
-      dF <- function(...) {data.frame(..., check.names = FALSE)}
+      dF <- function(...) {
+        data.frame(..., check.names = FALSE)
+      }
       if (is.null(d)) d <- getData(m, subset = TRUE, env = env)
       x <- dF(model.matrix(m, data = d))[en]
       inx <- any(isInx(en))  # interactions?